+++ src/main/java/org/kakaoshare/backend/domain/cart/controller/CartController.java
@@ -2,15 +2,12 @@ package org.kakaoshare.backend.domain.cart.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.kakaoshare.backend.domain.cart.dto.CartClearResponse;
-import org.kakaoshare.backend.domain.cart.dto.CartDeleteResponse;
-import org.kakaoshare.backend.domain.cart.dto.CartRegisterRequest;
-import org.kakaoshare.backend.domain.cart.dto.CartRegisterResponse;
-import org.kakaoshare.backend.domain.cart.dto.CartResponse;
-import org.kakaoshare.backend.domain.cart.entity.Cart;
+import org.kakaoshare.backend.domain.cart.dto.delete.CartClearResponse;
+import org.kakaoshare.backend.domain.cart.dto.delete.CartDeleteResponse;
+import org.kakaoshare.backend.domain.cart.dto.register.CartRegisterRequest;
+import org.kakaoshare.backend.domain.cart.dto.register.CartRegisterResponse;
+import org.kakaoshare.backend.domain.cart.dto.inquiry.CartResponse;
 import org.kakaoshare.backend.domain.cart.service.CartService;
-import org.kakaoshare.backend.domain.funding.dto.RegisterRequest;
-import org.kakaoshare.backend.domain.funding.dto.RegisterResponse;
 import org.kakaoshare.backend.jwt.util.LoggedInMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,7 +15,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
