+++ src/main/java/org/kakaoshare/backend/domain/cart/dto/delete/CartDeleteResponse.java
@@ -1,4 +1,4 @@
-package org.kakaoshare.backend.domain.cart.dto;
+package org.kakaoshare.backend.domain.cart.dto.delete;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import org.kakaoshare.backend.domain.cart.entity.Cart;
 
 @Getter
 @Builder
-public class CartRegisterResponse {
+public class CartDeleteResponse {
     private Long cartId;
 
-    public static CartRegisterResponse from(Cart cart){
-        return CartRegisterResponse.builder()
+    public static CartDeleteResponse from(Cart cart){
+        return CartDeleteResponse.builder()
                 .cartId(cart.getCartId())
                 .build();
     }
