+++ src/main/java/org/kakaoshare/backend/domain/cart/dto/register/CartRegisterRequest.java
@@ -1,4 +1,4 @@
-package org.kakaoshare.backend.domain.cart.dto;
+package org.kakaoshare.backend.domain.cart.dto.register;
 
 import lombok.Builder;
 import lombok.Getter;
